requirements.txt
APScheduler==3.6.3
certifi==2021.5.30
greenlet==1.1.0
psycopg2-binary==2.8.6
python-telegram-bot==13.5
pytz==2021.1
six==1.16.0
SQLAlchemy==1.4.17
tornado==6.1
tzlocal==2.1
